platform_adapter/requirements.txt
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0
